reshuffle directories
I am about to start the causal broadcast implementation and i'm trying
to move things around a bit. I want a good directory struture for this
project so that's why. The new causal_broadcast directory also contains
the implementation of the clock to be used to keep time of events that
will take place between nodes and in the cluster at large.

Let's get started.