Node for causal broadcast
Started the implementation for the node in a causal broadcast cluster.
I'm up late s theres a lot of filthy shoot and forget code in here that
does not work at all. But i took a jab at it is what matters. I'll look
at it at a later time when i'm free and fix the issues i have right now.
I'm exhausted and beat, i need some sleep

This commit contains changes to go.mod, i've included spew for logging
Added couple of methods to the clock package to make life easier
and did lots of things in the node package that do not work yet because
they are not well thought out.
Bye...