add event processing functionality to nodes
Nodes should be able to decide what to do with messages they recieve.
Not all messages have to be delivered to the node history, some have
to be queued to be delivered at a later time. And others have to be
dropped for reason i'll come to know later. For now, basic queueing,
checking for causal consistency of events is where we at.